bharat-context-adaptive-engine/requirements.txt
# Bharat Context-Adaptive Engine - Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# YAML parsing
pyyaml==6.0.1

# LLM & API
openai>=1.0.0
httpx==0.25.2
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
